git-svn-id: https://svnsrv.sdu.dk/svn/RobWork/trunk@2733 494c09c7-a67d-49d4-9938-151fb1965be8

+++ RobWork/src/rw/math/Q.cpp
@@ -86,7 +86,7 @@ std::istream& rw::math::operator>>(std::istream& in, Q& q) {
     if (ch1 != '}')
         RW_THROW("Content of input stream does not match format of Q");
 
-	if (size > -1 && res.size() != size) {
+	if (size > -1 && (int)res.size() != size) {
 		RW_THROW("Length of Q does not match device");
 	}
